# Decode WAL in Separate Goroutine