This PR moves the decoding of WAL records into a separate goroutine which passes the decoded results back using a channel. When looking at startup performance in a handful of Prometheus instances, startup time is reasonably evenly split between decoding records and processing the result so this has the potential to reduce WAL replay time by up to 50%.

Benchmarks show a reasonable performance improvement.

WAL compression off:
```
benchmark                                                                  old ns/op     new ns/op     delta
BenchmarkLoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200-8     442345548     390425328     -11.74%
BenchmarkLoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50-8     896853353     852906578     -4.90%
BenchmarkLoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480-8     303195989     275009282     -9.30%
```
WAL compression on:
```
benchmark                                                                  old ns/op     new ns/op     delta
BenchmarkLoadWAL/batches=10,seriesPerBatch=100,samplesPerSeries=7200-8     475893407     439551460     -7.64%
BenchmarkLoadWAL/batches=10,seriesPerBatch=10000,samplesPerSeries=50-8     997983033     852942115     -14.53%
BenchmarkLoadWAL/batches=10,seriesPerBatch=1000,samplesPerSeries=480-8     329612836     289762604     -12.09%
```

The first commit on this PR improves the benchmark for replaying the WAL, trying to use more reasonable values for the number of samples per series, and being sure to use `b.N` so the reporting is correct.