@csmarchbanks 2019-10-30
I am planning to let the WAL build to a reasonable size and the manually restart the Prometheus instances to see how long startup takes.