@csmarchbanks 2019-10-31
/benchmark cancel

Just restarted each Prometheus. Timing numbers are from the log line for starting to replay the WAL to the last segment loaded.
```
master       7m59s
pr           5m52s
improvement: 23.3%
```