@geekodour 2019-11-02
> we don't restart the Prometheus 

There's a half baked PR to prombench on restarting prometheus at random intervals, will try to add it soon.

cc @codesome