@cstyan 2019-11-04
Looks reasonable to me, I'll read through again later today. Is the memory usage when replaying any different with the use of the `sync.Pool`'s?