@csmarchbanks 2019-11-04
> s the memory usage when replaying any different with the use of the sync.Pool's?

I didn't notice much of a difference. The WAL is replayed faster, so memory growth is also faster as expected. The pool shouldn't contain more than 10 items, so it is pretty limited in size.